DEV: new irrigation method and synthetic exp.
1) Ensemble mean irrigation trigger (bug fix)
When the perturbation bias correction is turned off. Irrigation remains
triggered individually for each ensemble member.
When the perturbation bias correction is ON, the ensemble mean moisture
availability is computed, and the ensemble mean of the vegetation
trigger. This was done to avoid over irrigation due to the bias
correction.
Files changed: noahmp36_getirrigationstated.F90

2) Shift in precipitation time window (for synthetic experiment)
A new option is available in the config file:
"Shifted precipitation time window (MERRA2):"
where the shift (in years) can be inditated. E.g. -10 will take the
precipitation of the same date but 10 years ago.
This is only valid for precipitation and MERRA2 forcing.
Files changed: LIS_PRIV_rcMod.F90, LIS_readConfig.F90,
get_merra2.F90, read_merra2.F90.

3) New method for irrigation based on innovation jump detection
Irrigation (NoahMP3.6) can only be triggered when a high enough jump is
detected in the innovations and when the soil is not too wet. This
option is enabled with a new attribute in the config file:
"NEW_option:" (1 - enables, 0 - disabled)
In "enkf_Mod.F90" there is a check to see if the DELTAinnov (innovt -
innovt-1) is larger than the defined threshold in the config file:
"thresh4irr:"
If this is the case, an irrigation flag in NoahMP is turned on for all
ensemble members. Another check occurs for the available moisture, it
has to be inferior to a certain threshold defined in the config file:
"Irrigation threshold:"
The soil moisture incrmeents are not applied for an irrigation day.
Files changed: LIS_PRIV_rcMod.F90, LIS_irrigationMod.F90, enkf_Mod.F90,
NoahMP36_lsmMod.f90, NoahMP36_module.F90, noahmp36_updatesoilm.F90,
noahmp36_getirrigationstates.F90