ENH: add second threshold for irrigation BC
DA via the buddy check approach now allows two ways for irrigation
to be triggered.
  1. [ORIGINAL BC APPROACH] deltaInnov > 2.5 stdev and MA < MA_irr + 0.10
  2. [NEW] deltaInnov > stdev and MA < MA_irr

When the innovation is a clear outlier, the MA threshold allows more
flexibility (+0.10 = 10% of TAW).
When the outlier is not so trivial, the MA threshold is set to
the original MA_irr threshold.

Note: the buddy check approach should eventually be implemented in
the EnKF and not directly linked to the model. Tthe irrigation flags,
innovation arrays,... should be part of the enkf structure.