ENH: MA multiplier implementation
Removal of strict MA thresholds (incl. two trigger methods).
MA is now considered as a multiplier. The irrigation
check is now moved to the enkf_Mod.f90 which uses the stored
MA variable from Noah-MP.

This also corrects a previous bug where outliers were never
used to update the states, even if irrigation was not triggered.

To store the MA, the irrigation times and checks were changed,
future edits will deal with this in a more appropriate way.